Move helper methods to PatchworkMappingService
